backend/requirements.txt
# Flask, server
flask[async]>=3.0.3
flask-limiter>=3.5.0
flask-cors>=4.0.0
gunicorn>=22.0.0

# Database (Google Cloud Firestore)
google-cloud-firestore>=2.14.0

# GCP Secret Manager
google-cloud-secret-manager==2.18.3

# Validation and serialization
pydantic>=2.7.0
pydantic-settings>=2.3.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0

# Push notifications (APNs)
aioapns==3.1

# Push notifications (FCM / Android)
firebase-admin==6.4.0

# HTTP client for external APIs
httpx==0.26.0

# Machine learning for probability model
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4

# Background tasks
apscheduler==3.10.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
freezegun==1.4.0
# Utilities
python-dotenv==1.0.0

# Redis cache
redis[asyncio]>=5.0

# Anthropic VLM for ticket OCR
anthropic>=0.40.0

# Email (OTP delivery)
resend>=2.0.0

# Image processing (compress before sending to Anthropic API)
Pillow>=10.0.0
